requirements.txt
# Multi-Agent PowerPoint Orchestrator
# Core dependencies

# PowerPoint generation
python-pptx>=0.6.21

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# AI/LLM integration (for agent intelligence)
anthropic>=0.18.0

# Async support
asyncio-throttle>=1.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0  # For beautiful console output
